{"schema":"decodex/commit/1","summary":"Consolidate intake candidate selection test boundaries","authority":"manual"}
