--- apps/decodex/src/orchestrator/tests/intake_candidate_selection/closeout_policy/pr_state_errors.rs
@@ -1,55 +1,0 @@
-use crate::{
-	orchestrator::{
-		self, IssueDispatchMode, TargetIssueRunContext,
-		tests::{self, FakeTracker, intake_candidate_selection::support},
-	},
-	state::StateStore,
-	worktree::WorktreeManager,
-};
-
-#[test]
-fn non_dry_run_closeout_dispatch_errors_when_pr_state_read_fails() {
-	let (_temp_dir, base_config, workflow) = tests::temp_project_layout();
-	let config = tests::service_config_with_github_token_env_var(
-		&base_config,
-		"DECODEX_TEST_MISSING_DIRECT_DELIVERY_CLOSEOUT_GITHUB_TOKEN",
-	);
-	let issue = support::candidate_selection_service_owned_issue("Done");
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let worktree_manager =
-		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
-	let worktree =
-		worktree_manager.ensure_worktree(&issue.identifier, false).expect("worktree should exist");
-	let head_oid = tests::git_output(&worktree.path, &["rev-parse", "HEAD"]);
-	let pr_url = "https://github.com/hack-ink/decodex/pull/179";
-
-	tests::seed_review_handoff_marker_value(
-		&state_store,
-		config.service_id(),
-		&issue.id,
-		&tests::sample_review_handoff_marker(&worktree.branch_name, pr_url, &head_oid),
-	);
-
-	let error = orchestrator::run_target_issue_once(TargetIssueRunContext {
-		tracker: &tracker,
-		project: &config,
-		workflow: &workflow,
-		state_store: &state_store,
-		issue_id: &issue.id,
-		preferred_issue_state: None,
-		preferred_initial_issue_state: Some("In Review"),
-		dry_run: false,
-		lease_preacquired: false,
-		preferred_issue_claim_fd: None,
-		preferred_dispatch_slot_fd: None,
-		preferred_dispatch_slot_index: None,
-		dispatch_mode: IssueDispatchMode::Closeout,
-		preferred_run_identity: None,
-		preferred_retry_budget_base: None,
-	})
-	.expect_err("non-dry-run closeout dispatch should surface GH state read failures");
-
-	assert!(error.to_string().contains("pull_request_state_read_failed"));
-}
