--- apps/decodex/src/orchestrator/tests/intake_candidate_selection/post_review_repair/targeted_retry_blocks_retained_review_handoff_marker_in_state_transition_window.rs
@@ -1,65 +1,0 @@
-use crate::{
-	orchestrator::{
-		self, IssueDispatchMode, TargetIssueRunContext,
-		tests::{self, FakeTracker, TEST_SERVICE_ID},
-	},
-	state::StateStore,
-	tracker,
-	worktree::WorktreeManager,
-};
-
-#[test]
-fn targeted_retry_blocks_retained_review_handoff_marker_in_state_transition_window() {
-	let (_temp_dir, config, workflow) = tests::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let active_label = tracker::automation_active_label(TEST_SERVICE_ID);
-	let issue = tests::sample_issue("In Progress", &[active_label.as_str()]);
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let worktree_manager =
-		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
-	let worktree =
-		worktree_manager.ensure_worktree(&issue.identifier, false).expect("worktree should exist");
-	let head_oid = tests::git_output(&worktree.path, &["rev-parse", "HEAD"]);
-	let pr_url = "https://github.com/hack-ink/decodex/pull/174";
-
-	state_store
-		.upsert_worktree(
-			config.service_id(),
-			&issue.id,
-			&worktree.branch_name,
-			&worktree.path.display().to_string(),
-		)
-		.expect("worktree should record");
-
-	tests::seed_review_handoff_marker_value(
-		&state_store,
-		config.service_id(),
-		&issue.id,
-		&tests::sample_review_handoff_marker(&worktree.branch_name, pr_url, &head_oid),
-	);
-
-	let summary = orchestrator::run_target_issue_once(TargetIssueRunContext {
-		tracker: &tracker,
-		project: &config,
-		workflow: &workflow,
-		state_store: &state_store,
-		issue_id: &issue.id,
-		preferred_issue_state: None,
-		preferred_initial_issue_state: None,
-		dry_run: true,
-		lease_preacquired: false,
-		preferred_issue_claim_fd: None,
-		preferred_dispatch_slot_fd: None,
-		preferred_dispatch_slot_index: None,
-		dispatch_mode: IssueDispatchMode::Retry,
-		preferred_run_identity: None,
-		preferred_retry_budget_base: None,
-	})
-	.expect("targeted retry planning should succeed");
-
-	assert!(
-		summary.is_none(),
-		"retry dispatch must not mint a duplicate attempt for a retained review handoff lane"
-	);
-}
